[BE] Refactor: 리뷰 사항 수정
- dao 에서 repository 로 롤백
- NullPointerException 롤백
