+++ BE/src/main/java/com/codesquad/sidedish08/message/ErrorMessages.java
@@ -4,5 +4,4 @@ public class ErrorMessages {
 
   public static final String ERROR = "ERROR";
   public static final String NO_SUCH_ELEMENT_EXCEPTION = "요청 값의 객체를 찾을 수 없습니다";
-  public static final String NULL_POINTER_EXCEPTION = "NULL POINTER EXCEPTION";
 }
