+++ BE/src/main/java/com/codesquad/sidedish08/repository/CategoryRepository.java
@@ -1,4 +1,4 @@
-package com.codesquad.sidedish08.dao;
+package com.codesquad.sidedish08.repository;
 
 
 import com.codesquad.sidedish08.model.Category;
