+++ BE/src/main/java/com/codesquad/sidedish08/service/MainDishService.java
@@ -1,5 +1,6 @@
 package com.codesquad.sidedish08.service;
 
+
 import static com.codesquad.sidedish08.util.ResponseUtils.getResultMap;
 
 import com.codesquad.sidedish08.repository.CategoryRepository;
@@ -25,7 +26,7 @@ public class MainDishService implements BasicService {
     return getResultMap("body", jdbcCategoryRepository.findById());
   }
 
-  public Map<String, ?> detail(String detailHash) {
+  public Map<String, Object> detail(String detailHash) {
     return getResultMap("dish", null);
   }
 }
